CAMEL-19922 - Fix storing escaped characters of Java Properties file
Previously, it was writing in the properties file directly without using
the Java Properties API. Consequently, the special characters are not
encoded. In spotted issue for Windows path, the backslash. When reading
back from it, we are using the Java Properties API which is removing the
special character backslash.

Using the Java Properties API also to write the data is properly
escaping/de-escaping characters.

The test is modified to use files allowing to reproduce the initially
reported issue on Windows.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>