CAMEL-2219 - Include API methods without parameter in catalog
- the api methods were added when the parameters, also named
apiProperties but when there were none of them, the api methods was not
added
- here is a proposal to add all api methods first and in a single place
- it is also adding empty list of apiProperties entries. it has the
advantage that when we search for available apiProperties for an api
method there is an empty list instead of a null returned and so will
avoid potential breakage in consumer of this API, and an easier way to
handle it. On the other hand, it is adding a bit more lines in the
catalog which are not bringing new information.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>